[#1]refactor(dto): DTO 내 List 속성은 Inner Class로 옮김
