--- src/main/java/baksakcci/shoppingmall/order/application/dto/OrderItemCreate.java
@@ -1,12 +1,0 @@
-package baksakcci.shoppingmall.order.application.dto;
-
-import lombok.Builder;
-import lombok.Getter;
-
-@Builder
-@Getter
-public class OrderItemCreate {
-
-    private Long productId;
-    private int qty;
-}
