+++ src/main/java/baksakcci/shoppingmall/order/presentation/request/OrderRequest.java
@@ -1,8 +1,9 @@
 package baksakcci.shoppingmall.order.presentation.request;
 
 import baksakcci.shoppingmall.order.application.dto.OrderCreate;
-import baksakcci.shoppingmall.order.application.dto.OrderItemCreate;
+import baksakcci.shoppingmall.order.application.dto.OrderCreate.OrderItemCreate;
 import java.util.List;
+import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 
@@ -18,7 +19,7 @@ public class OrderRequest {
 
     public static OrderCreate toDto(Long userId, OrderRequest orderRequest) {
         List<OrderItemCreate> orderItemCreateDtos = orderRequest.orderItemRequests.stream()
-                .map(each -> OrderItemCreate.builder()
+                .map(each -> OrderCreate.OrderItemCreate.builder()
                         .productId(each.getProductId())
                         .qty(each.getQty())
                         .build())
@@ -32,5 +33,14 @@ public class OrderRequest {
                 .detailAddress(orderRequest.detailAddress)
                 .build();
 
+    }
+
+    @AllArgsConstructor
+    @Getter
+    @Builder
+    public static class OrderItemRequest {
+
+        private int qty;
+        private long productId;
     }
 }
