+++ src/test/java/baksakcci/shoppingmall/order/presentation/PlaceOrderControllerTest.java
@@ -4,7 +4,6 @@ import static org.springframework.test.web.servlet.request.MockMvcRequestBuilder
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-import baksakcci.shoppingmall.order.presentation.request.OrderItemRequest;
 import baksakcci.shoppingmall.order.presentation.request.OrderRequest;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import java.util.ArrayList;
